requirements.txt
asgiref==3.6.0
Django==4.2
django-recaptcha==3.0.0
tzdata==2023.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
requests
webdriver-manager
urllib3
bs4
colorama
regex
python-csv